requirements.txt
discord
python-dotenv
pymongo
apscheduler
flask